RunApplication Updated to support redrawing form anywhere.
when user selects pockets, the beads animate as needed. still working on asynchronous movement. other plays such as capture and collection of side at end of game are not updated to support animation yet. Added class to pass information through to redraw from animate. Do not change the flow of information from the thread call to the animateBeads method.